backend/requirements.txt
boto3
chalice
PyJWT
cryptography
requests
